refactor: 중복된 ParseInt check 삭제 및 utils클래스로 병합
