--- src/main/java/step2/validator/Validator.java
@@ -1,6 +1,0 @@
-package step2.validator;
-
-public interface Validator {
-
-    boolean validate(String[] sources);
-}
